+++ frontend/components/page/head.php
@@ -1,7 +1,5 @@
+<!-- TODO: implement get_page_title with this fallback -->
 <?php if (!isset($title)) $title = 'Pepe Fabra Valverde - M.L. & A.I. Portfolio' ?>
-<?php $default_styles = [ 'main', 'header', 'footer' ] ?>
-
-<?php $styles = array_unique( array_merge($styles, $default_styles) ) ?>
 
 <head>
     <meta http-equiv="Content-Type" content="text/html;" charset="utf-8" /> 
@@ -10,18 +8,4 @@
     <meta http-equiv="Content-Security-Policy" content="default-src 'self'; img-src https://*; child-src 'none';">
 
     <?= c('page/seo', [ 'title' => $title ]) ?>
-    <style>
-        <?php foreach ($styles as $style): ?>
-            <?php
-                // TODO: implement function
-                $full_path = path_join(PUBLIC_DIR, 'styles', "$style.css");
-
-                $content = '';
-                if (file_exists($full_path)) $content = file_get_contents($full_path);
-            ?>
-            <?= $content ?>
-        <?php endforeach; ?>
-
-        /* TODO: implement save it in a separate file, and probably later minify it */
-    </style>
 </head>